Extract ICS export into testable module with unit tests
- Move exportICS() from Workspace.tsx to src/lib/ics.ts
- Add 7 unit tests covering: VCALENDAR structure, single/multi-day
  events, disabled courses, TBA times, multiline times, section
  selection
- Remove unused TERM_START_DATES import from Workspace.tsx
